+++ backend/app/cms/main.py
@@ -1,5 +1,4 @@
 import uuid
-from typing import List
 from fastapi import APIRouter, Depends, HTTPException, status, Query
 from sqlmodel import Session, select
 from sqlalchemy.exc import MultipleResultsFound, NoResultFound
@@ -7,11 +6,11 @@ from sqlalchemy.sql import literal_column
 import logging
 
 from app.core.db import engine
-from app.models import (
+from app.cms.models import (
     CMSContentCreate,
     CmsContentUpdate,
-    CMSContentPublicWithLanguages,
-    AllCMSContentPublic,
+    CMSContentPublish,
+    CMSContentDelete,
     LanguageEnum,
     CMS_MODEL_MAP,
     CMSContentTypesEnum,
@@ -27,19 +26,18 @@ def get_session():
 
 
 @router.post(
-    "/content/{type}",
-    response_model=CMSContentCreate,
+    "/content",
+    # response_model=CMSContentCreate,
     status_code=status.HTTP_201_CREATED,
 )
 async def create_cms_content(
-    type: CMSContentTypesEnum,
     data: CMSContentCreate,
     session: Session = Depends(get_session),
 ):
     """Create a new CMS content entry"""
     # Check if content with same slug and language already exists
-    CmsModel = CMS_MODEL_MAP[type]
-    print("!!!", data)
+    CmsModel = CMS_MODEL_MAP[data.content_type]
+
     existing = session.exec(
         select(CmsModel)
         .where(CmsModel.slug == data.slug)
@@ -60,21 +58,127 @@ async def create_cms_content(
     return content
 
 
+@router.patch("/content", 
+            #   response_model=AllCMSContentPublic
+              )
+async def update_cms_content(
+    data: CmsContentUpdate,
+    session: Session = Depends(get_session),
+):
+    """Update existing CMS content"""
+    CMSModel = CMS_MODEL_MAP[data.content_type]
+    # Check if content exists
+    content = session.exec(
+        select(CMSModel).where(CMSModel.id == data.content_id)
+    ).first()
+
+    if not content:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Content with ID '{data.content_id}' not found",
+        )
+
+    # If updating slug or language, check for conflicts
+    if (data.slug and data.slug != content.slug) or (
+        data.language and data.language != content.language
+    ):
+        new_slug = data.slug or content.slug
+        new_language = data.language or content.language
+
+        conflict = session.exec(
+            select(CMSModel)
+            .where(CMSModel.slug == new_slug)
+            .where(CMSModel.language == new_language)
+            .where(CMSModel.id != data.content_id)
+        ).first()
+
+        if conflict:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail=f"Content with slug '{new_slug}' and language '{new_language}' already exists",
+            )
+
+    # Update content
+    update_data = data.model_dump(exclude_unset=True)
+    for key, value in update_data.items():
+        setattr(content, key, value)
+
+    session.add(content)
+    session.commit()
+    session.refresh(content)
+
+    return content
+
+
+@router.post("/content/publish", 
+            #  response_model=AllCMSContentPublic
+             )
+async def publish_cms_content(
+
+    data: CMSContentPublish ,session: Session = Depends(get_session)):
+    """Publish draft content"""
+    CMSModel = CMS_MODEL_MAP[data.content_type]
+    content = session.exec(
+        select(CMSModel).where(CMSModel.id == data.content_id)
+    ).first()
+
+    if not content:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Content with ID '{data.content_id}' not found",
+        )
+
+    if not content.draft_content:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="No draft content to publish",
+        )
+
+    content.published_content = content.draft_content
+    content.draft_content = None
+    session.add(content)
+    session.commit()
+    session.refresh(content)
+
+    return content
+
+
+@router.delete("/content", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_cms_content(data: CMSContentDelete, session: Session = Depends(get_session)):
+    """Delete CMS content by ID"""
+    logger.info("DELETING ", data)
+    CMSModel = CMS_MODEL_MAP[data.content_type]
+    content = session.exec(
+        select(CMSModel).where(CMSModel.id == data.content_id)
+    ).first()
+
+    if not content:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Content with ID '{data.content_id}' not found",
+        )
+
+    session.delete(content)
+    session.commit()
+
+    return None
+
+
 @router.get(
-    "/content/{type}",
+    "/content/{content_type}/list",
     # response_model=list[AllCMSContentPublic]
 )
 async def list_cms_content(
-    type: CMSContentTypesEnum,
+    content_type: CMSContentTypesEnum,
     language: LanguageEnum = None,
-    districtr_map_slug: str = None,
-    districtr_map_slugs: List[str] = None,
+    # districtr_map_slug: str = None,
+    # districtr_map_slugs: List[str] = None,
     session: Session = Depends(get_session),
     offset: int = Query(default=0, ge=0),
     limit: int = Query(default=50, le=100),
 ):
     """List CMS content with optional filtering"""
-    CMSModel = CMS_MODEL_MAP[type]
+    CMSModel = CMS_MODEL_MAP[content_type]
     query = select(CMSModel)
 
     if language:
@@ -95,16 +199,18 @@ async def list_cms_content(
     return results
 
 
-@router.get("/content/{type}/{slug}", response_model=CMSContentPublicWithLanguages)
+@router.get(
+    "/content/{content_type}/slug/{slug}"
+    # response_model=CMSContentPublicWithLanguages
+)
 async def get_cms_content(
-    type: CMSContentTypesEnum,
+    content_type: CMSContentTypesEnum,
     slug: str,
     language: LanguageEnum = LanguageEnum.ENGLISH,
     session: Session = Depends(get_session),
 ):
     """Get CMS content by slug and language"""
-
-    CmsModel = CMS_MODEL_MAP[type]
+    CmsModel = CMS_MODEL_MAP[content_type]
 
     try:
         content = session.exec(select(CmsModel).where(CmsModel.slug == slug)).all()
@@ -130,102 +236,3 @@ async def get_cms_content(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail=f"Multiple content entries found for slug '{slug}' and language '{language}'",
         )
-
-
-@router.put("/content/{type}/{content_id}", response_model=AllCMSContentPublic)
-async def update_cms_content(
-    type: CMSContentTypesEnum,
-    content_id: str,
-    data: CmsContentUpdate,
-    session: Session = Depends(get_session),
-):
-    """Update existing CMS content"""
-    CMSModel = CMS_MODEL_MAP[type]
-    # Check if content exists
-    content = session.exec(select(CMSModel).where(CMSModel.id == content_id)).first()
-
-    if not content:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Content with ID '{content_id}' not found",
-        )
-
-    # If updating slug or language, check for conflicts
-    if (data.slug and data.slug != content.slug) or (
-        data.language and data.language != content.language
-    ):
-        new_slug = data.slug or content.slug
-        new_language = data.language or content.language
-
-        conflict = session.exec(
-            select(CMSModel)
-            .where(CMSModel.slug == new_slug)
-            .where(CMSModel.language == new_language)
-            .where(CMSModel.id != content_id)
-        ).first()
-
-        if conflict:
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail=f"Content with slug '{new_slug}' and language '{new_language}' already exists",
-            )
-
-    # Update content
-    update_data = data.model_dump(exclude_unset=True)
-    for key, value in update_data.items():
-        setattr(content, key, value)
-
-    session.add(content)
-    session.commit()
-    session.refresh(content)
-
-    return content
-
-
-@router.delete("/content/{type}/{content_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_cms_content(
-    content_id: str, type: CMSContentTypesEnum, session: Session = Depends(get_session)
-):
-    """Delete CMS content by ID"""
-    CMSModel = CMS_MODEL_MAP[type]
-    content = session.exec(select(CMSModel).where(CMSModel.id == content_id)).first()
-
-    if not content:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Content with ID '{content_id}' not found",
-        )
-
-    session.delete(content)
-    session.commit()
-
-    return None
-
-
-@router.post("/content/{type}/{content_id}/publish", response_model=AllCMSContentPublic)
-async def publish_cms_content(
-    content_id: str, type: CMSContentTypesEnum, session: Session = Depends(get_session)
-):
-    """Publish draft content"""
-    CMSModel = CMS_MODEL_MAP[type]
-    content = session.exec(select(CMSModel).where(CMSModel.id == content_id)).first()
-
-    if not content:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"Content with ID '{content_id}' not found",
-        )
-
-    if not content.draft_content:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="No draft content to publish",
-        )
-
-    content.published_content = content.draft_content
-    content.draft_content = None
-    session.add(content)
-    session.commit()
-    session.refresh(content)
-
-    return content
